MR-1003 TS: ErrorOverlay + FloorAndTagControls + FloorLabel
